doc_extraction/requirements.txt
google-genai
python-dotenv
pydantic
PyMuPDF
